pyproject.toml
[project]
name = "ship-sat"
version = "0.1.0"
description = "Ship detection on satellite imagery: Airbus RLE to YOLO-OBB, Ultralytics training, SAHI inference, Streamlit UI."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ultralytics>=8.3.0",
    "sahi>=0.11.18",
    "streamlit>=1.28.0",
    "opencv-python-headless>=4.8.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "kaggle>=1.6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0"]
viz = ["matplotlib>=3.8.0"]

[project.scripts]
ship-sat-download = "ship_sat.download:main"
ship-sat-prep = "ship_sat.dataset_prep:main"
ship-sat-train = "ship_sat.train:main"
ship-sat-infer = "ship_sat.infer_sahi:main"

[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ship_sat"]

[tool.hatch.build.targets.sdist]
include = ["src/ship_sat", "README.md", "configs"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
